ssr: get ssr worked
