+++ .github/workflows/main.yml
@@ -2,9 +2,9 @@ name: CI
 
 on:
   push:
-    branches: [ master ]
+    branches: [ ssr-experiment ]
   pull_request:
-    branches: [ master ]
+    branches: [ ssr-experiment ]
 
 jobs:
   build:
