+++ app/db/files.ts
@@ -1,7 +1,8 @@
-import { extractFileDetail, FileData, FileDetail, files } from "../db";
-import { generateRandomBytes, Result } from "../utils";
+import { extractFileDetail, files } from '../db';
+import { generateRandomBytes, Result } from '../utils';
 import { promises as fs } from 'fs';
-import { FILE_NOT_FOUND } from "../errors";
+import { FILE_NOT_FOUND } from '../errors';
+import { FileData, FileDetail } from 'app/types';
 
 async function getNewFileId() {
   let fid = generateRandomBytes(16);
