+++ public/index.html
@@ -3,18 +3,13 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width,initial-scale=1">
-  <!--vue-meta-outlet-->
   <script src="https://cdn.jsdelivr.net/npm/vue@3.0.2/dist/vue.global.prod.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/vue-router@4.0.0-rc.5/dist/vue-router.global.prod.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/vuex@4.0.0-rc.1/dist/vuex.global.prod.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/element-plus@1.0.1-beta.0/lib/index.full.js"></script>
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/element-plus@1.0.1-beta.0/lib/theme-chalk/index.css">
-  <link rel="stylesheet" href="/style.css">
 </head>
 <body>
-  <div id="app">
-    <!--vue-ssr-outlet-->
-  </div>
-  <script src="/main.js"></script>
+  <div id="app"></div>
 </body>
 </html>
