+++ tsconfig.json
@@ -14,9 +14,5 @@
         "src/*"
       ]
     }
-  },
-  "include": [
-    "app/**/*.ts",
-    "src/**/*.ts",
-  ]
+  }
 }
